Split PluginInstanceHostAudioThread in its own file
